requirements.txt
anyio==4.2.0
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
python-telegram-bot==13.9
pytz==2024.1
six==1.16.0
sniffio==1.3.0
tornado==6.4
typing_extensions==4.9.0
tzlocal==5.2